Add google-cloud-text_to_speech and figaro gems, create GoogleTextToSpeechService
Added "google-cloud-text_to_speech" and "figaro" gems to the Gemfile to enable Text-to-Speech functionality using Google Cloud's Text-to-Speech service. The figaro gem was added for secure configuration of sensitive data.

Also, a new service “GoogleTextToSpeechService” has been implemented to synthesize speech, leveraging Google Cloud's Text-to-Speech utility. This addition is a part of the initiative to implement a feature for translating text inputs into speech.