Add audio synthesis feature and related specs
Implemented the audio synthesis feature, which allows system to synthesize audio from the subject body text and attach it to the subject. This includes necessary interactors for synthesizing audio, managing audio synthesis for a subject, and saving audio to the subject. Also added relative spec files for each of these interactors.

Additional changes include the addition of various gems required for this feature such as ‘vcr’, 'webmock', 'activeinteractor', 'pry' and updates in rails_helper, Gemfile and test environment configurations.

This feature gives system the capability to generate audio from text, which might enhance the user experience by providing an alternative, auditory way to digest information.